feat(cli): Update CLI to support source v1 protocol
